Crate canopy_widgets

Built-in widgets for canopy applications.

Re-exports

pub use editor::Editor;

Modules

editor

Editor widget.

frame

Frame widget.

inspector

Inspector widget.

list

List widget.

tabs

Tabs widget.

Structs

Input

A single input line, one character high. Single-line text input widget.

Panes

Panes manages a set of child nodes arranged in a 2d grid.

Root

A Root node that lives at the base of a Canopy app. It manages modal windows, houses the Inspector and exposes a set of built-in functions.

Text

Multiline text widget with wrapping and scrolling.