canopy_widgets::editor

Trait Pos

pub trait Pos: Sized {
    // Required methods
    fn new(s: &State, chunk: usize, offset: usize) -> Self;
    fn constrain(&self, s: &State) -> Self;
    fn chunk_offset(&self) -> (usize, usize);
    fn is_between(&self, s: &State) -> bool;

    // Provided methods
    fn shift(&self, s: &State, n: isize) -> Self { ... }
    fn shift_chunk(&self, s: &State, n: isize) -> Self { ... }
    fn shift_line(&self, s: &State, n: isize) -> Self { ... }
}

A position that can be clamped within the bounds of a State.

Required Methods

fn new(s: &State, chunk: usize, offset: usize) -> Self

Create a new item and clamp it

fn constrain(&self, s: &State) -> Self

Constrain within state bounds, and return a new item

fn chunk_offset(&self) -> (usize, usize)

Retrieve the chunk and offset of the cursor.

fn is_between(&self, s: &State) -> bool

Is this cursor between wrapped lines?

Provided Methods

fn shift(&self, s: &State, n: isize) -> Self

Shift the cursor by an offset within a chunk. If the new position is out of bounds, return the closest matching position within the chunk. If the new offset lands on a character that is between lines, we continue in the same direction until we find a character that is in bounds.

fn shift_chunk(&self, s: &State, n: isize) -> Self

Shift the chunk offset. If the new position is out of bounds, return the closest matching position.

fn shift_line(&self, s: &State, n: isize) -> Self

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.