canopy_widgets

Struct Panes

pub struct Panes { /* private fields */ }

Panes manages a set of child nodes arranged in a 2d grid.

Implementations

impl Panes

pub fn new() -> Self

Construct panes with no children.

pub fn with_child(child: NodeId) -> Self

Construct panes with a single child.

pub fn column_nodes(&self) -> Vec<NodeId>

Return the active column container node IDs in order.

pub fn focused_column_index(&self, c: &dyn Context) -> Option<usize>

Return the focused column index, if any.

pub fn focus_column(&mut self, c: &mut dyn Context, delta: i32) -> Result<()>

Move focus by a signed column offset (wraps around).

pub fn next_column(&mut self, c: &mut dyn Context) -> Result<()>

Move focus to the next column.

pub fn prev_column(&mut self, c: &mut dyn Context) -> Result<()>

Move focus to the previous column.

pub fn focus_coords(&self, c: &dyn Context) -> Option<(usize, usize)>

Get the offset of the current focus in the children vector.

pub fn delete_focus(&mut self, c: &mut dyn Context) -> Result<()>

Delete the focus node. If a column ends up empty, it is removed.

pub fn insert_row(&mut self, c: &mut dyn Context, n: NodeId) -> Result<()>

Insert a node, splitting vertically.

pub fn insert_col(&mut self, c: &mut dyn Context, n: NodeId) -> Result<()>

Insert a node in a new column.

pub fn cmd_focus_column() -> &'static CommandSpec

Return a typed command reference for this command.

pub fn cmd_next_column() -> &'static CommandSpec

Return a typed command reference for this command.

pub fn cmd_prev_column() -> &'static CommandSpec

Return a typed command reference for this command.