canopy_widgets

Struct Selector

pub struct Selector<T>
where
    T: SelectorItem,
{ /* private fields */ }

A multi-select widget with checkbox-style items.

Items can be toggled on/off independently. The selected indices are tracked in the order they were selected, allowing for ordered selection if needed.

Implementations

impl<T> Selector<T>
where T: SelectorItem + 'static,

pub fn new(items: Vec<T>) -> Self

Create a new selector with the given items.

pub fn selected_indices(&self) -> &[usize]

Get the selected indices in selection order.

pub fn selected_items(&self) -> Vec<&T>

Get references to the selected items in selection order.

pub fn is_selected(&self, index: usize) -> bool

Check if an index is selected.

pub fn focused_index(&self) -> usize

Get the currently focused index.

pub fn toggle(&mut self, _c: &mut dyn Context) -> Result<()>

Toggle selection of the focused item.

pub fn select_by(&mut self, _c: &mut dyn Context, delta: i32) -> Result<()>

Move focus by a signed offset.

pub fn select_next(&mut self, c: &mut dyn Context) -> Result<()>

Move focus to the next item.

pub fn select_prev(&mut self, c: &mut dyn Context) -> Result<()>

Move focus to the previous item.

pub fn select_first(&mut self, _c: &mut dyn Context) -> Result<()>

Move focus to the first item.

pub fn select_last(&mut self, _c: &mut dyn Context) -> Result<()>

Move focus to the last item.

pub fn clear(&mut self, _c: &mut dyn Context) -> Result<()>

Clear all selections.

pub fn select_all(&mut self, _c: &mut dyn Context) -> Result<()>

Select all items.

pub fn len(&self) -> usize

Get the number of items.

pub fn is_empty(&self) -> bool

Check if the selector is empty.

pub fn selected_count(&self) -> usize

Get the number of selected items.

pub fn cmd_toggle() -> &'static CommandSpec

Return a typed command reference for this command.

pub fn cmd_select_by() -> &'static CommandSpec

Return a typed command reference for this command.

pub fn cmd_select_next() -> &'static CommandSpec

Return a typed command reference for this command.

pub fn cmd_select_prev() -> &'static CommandSpec

Return a typed command reference for this command.

pub fn cmd_select_first() -> &'static CommandSpec

Return a typed command reference for this command.

pub fn cmd_select_last() -> &'static CommandSpec

Return a typed command reference for this command.

pub fn cmd_clear() -> &'static CommandSpec

Return a typed command reference for this command.

pub fn cmd_select_all() -> &'static CommandSpec

Return a typed command reference for this command.